LayoutTests/media/modern-media-controls/pip-support/pip-support-disabled-video-with-audio-tracks-only-expected.txt

Testing that the picture-in-picture button is disabled for <video> with an audio resource.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Once media loads its metadata the picture-in-picture button should become disabled
Obtained loadedmetadata event
PASS mediaController.isAudio is true
PASS mediaController.controls.pipButton.enabled is false

PASS successfullyParsed is true

TEST COMPLETE
